allow plotting
