+++ Cargo.toml
@@ -14,10 +14,10 @@ skyline-web = { git = "https://github.com/skyline-rs/skyline-web.git" }
 bitflags = "1.2.1"
 parking_lot = { version = "0.11.1", features = ["nightly"] }
 lazy_static = "1.4.0"
-owo-colors = "1.1.3"
-ramhorns = "0.10.1"
+owo-colors = "2.1.0"
+ramhorns = "0.12.0"
 paste = "1.0"
-num = "0.3.0"
+num = "0.4.0"
 num-derive = "0.3"
 num-traits = "0.2"
 wsl = "0.1.0"
@@ -26,6 +26,7 @@ strum_macros = "0.21.0"
 minreq = { version = "=2.2.1", features = ["https", "json-using-serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+training_mod_consts = { path = "training_mod_consts" }
 
 [patch.crates-io]
 ring = { git = "https://github.com/skyline-rs/ring", branch = "0.16.20" }
