+++ training_mod_consts/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "training_mod_consts"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+bitflags = "1.2.1"
+strum = "0.21.0"
+strum_macros = "0.21.0"
+num = "0.4.0"
+num-derive = "0.3"
+num-traits = "0.2"
+skyline_smash = { git = "https://github.com/ultimate-research/skyline-smash.git" }
